accomidate hardlinks properly

+++ lib/omnibus/file_syncer.rb
@@ -101,16 +101,28 @@ module Omnibus
             FileUtils.ln_sf(target, "#{destination}/#{relative_path}")
           end
         when :file
-          # First attempt a regular copy. If we don't have write
-          # permission on the File, open will probably fail with
-          # EACCES (making it hard to sync files with permission
-          # r--r--r--). Rescue this error and use cp_r's
-          # :remove_destination option.
-          begin
-            FileUtils.cp(source_file, "#{destination}/#{relative_path}")
-          rescue Errno::EACCES
-            FileUtils.cp_r(source_file, "#{destination}/#{relative_path}",
-                           :remove_destination => true)
+          source_stat = File.stat(source_file)
+          # Detect 'files' which are hard links and use ln instead of cp to
+          # duplicate them, provided their source is in place already
+          if hardlink? source_stat
+            if existing = hardlink_sources[[source_stat.dev, source_stat.ino]]
+              FileUtils.ln(existing, "#{destination}/#{relative_path}")
+            else
+              FileUtils.cp(source_file, "#{destination}/#{relative_path}")
+              hardlink_sources.store([source_stat.dev, source_stat.ino], "#{destination}/#{relative_path}")
+            end
+          else
+            # First attempt a regular copy. If we don't have write
+            # permission on the File, open will probably fail with
+            # EACCES (making it hard to sync files with permission
+            # r--r--r--). Rescue this error and use cp_r's
+            # :remove_destination option.
+            begin
+              FileUtils.cp(source_file, "#{destination}/#{relative_path}")
+            rescue Errno::EACCES
+              FileUtils.cp_r(source_file, "#{destination}/#{relative_path}",
+                             :remove_destination => true)
+            end
           end
         else
           raise RuntimeError,
@@ -154,6 +166,35 @@ module Omnibus
     #
     def relative_path_for(path, parent)
       Pathname.new(path).relative_path_from(Pathname.new(parent)).to_s
+    end
+
+    #
+    # A list of hard link file(s) sources which have already been copied,
+    # indexed by device and inode number.
+    #
+    # @api private
+    #
+    # @return [Hash{Array<FixNum, FixNum> => String}]
+    #
+    def hardlink_sources
+      @hardlink_sources ||= {}
+    end
+
+    #
+    # Determines whether or not a file is a hardlink.
+    #
+    # @param [File::Stat] stat
+    #   the File::Stat object for a file you wand to test
+    #
+    # @return [true, false]
+    #
+    def hardlink?(stat)
+      case stat.ftype.to_sym
+      when :file
+        stat.nlink > 1
+      else
+        false
+      end
     end
   end
 end
